"Semantically-type" output pins.
Instead of telling HAL to write to "pin X", we now tell HAL to write to
the BLOWER/SOLENOID pin.

This consolidates all of the "device X is attached to pin Y" logic into
HAL, so that we can handle a different pinout on e.g. STM32.

It also consolidates the setting of pins to OUTPUT into HAL.  We had a
bug a few days ago where we were forgetting to set the blower to OUTPUT.
But now HAL "owns" the notion that this pin is an output pin, so can
reasonably do this for us, in a consolidated place.

Also strongly-type the difference between "binary pins" that are always
either high or low (e.g. our solenoid) and pins that we pulse-width
modulate.  Of course any pin we PWM could also be set to all-high or
all-low, but in practice we treat them differently and don't want to mix
them up.

git-pr-chain: hal-output-pins