[GUI] Add script to run gui on eglfs 1920 x 1080 @ 60 mode
This script run_eglfs.sh is added alongside a kms configuration
file to force egl to run in a specific resolution/refresh_rate.

The script also enable some Qt debugging information to make it
easy to troubleshoot eventual issues with graphics

Also added a small fps counter element so we can keep track of
performance degratations. It will be available only in debug
builds once we start making this distinction.