[GUI] Add Style singleton and extract left-side-panel controls into reusable Item
On main.qml we had few hard-coded controls for "RR", "PEEP",
"PIP" and so on. This change extracts those into a reusable
element called StepCounter which allows user to increase
or decrease a value, one step at the time.

Aside from that, a singleton element Style is created
to hold all UI related configuration, such as themes,
fonts, margin sizes and so on.