[GUI] Update design of the graphs
Update design of the graphs for follow new recommendations.

Current implementation of the ScopeView has a very high cpu usage
making it unable to hold updates at 60Hz. (Actually I can at 90%
cpu usage).

Main reason is the fact we are using two Charts, one To draw a line
and the other to draw the area. Hopefully with enough time we can
customize the graphs to make it more efficient.