Send flow correction ml/s to GUI.
Will be used for disconnect detection.

I observe with the fan that a leak is about -1600ml/s.  Even when I set
PIP/PEEP to 5, we still open up the valve all the way.

I think it would be safe to set the disconnect detection threshold
around -500ml/s.

GPC: flow-correction-in-proto